Training in policy
For some algorithms (e.g VAE), some internal states (noise in VAE) need to be maintained for training. It's easier to directly
do the training in policy as we donot need to care about how to transport those internal states through replay buffer
to agent.train().

This PR implements a simple actor critic algorithm in policy.