squared root for distance in TracAlgorithm
The reason of using sqrt is that the change with sqrt is approximately proportional
to the step size so that the reducing the step by the ratio max_change/actual_change makes
more sense.

Also fixed a bug of TracAlgorithm when continuing training from restored checkpoint.
Previously, TrustedUpdater is created in rollout(), which is first called in
OnPolicyDriver/OffPolicyAlgorithm before the checkpoint is restored in PolicyTrainer.
So the prev_variables in TrustedUpdater stores randomly initialized varaibles instead
of previouly trained variables. This make the adjust_step() fail.

So move the creation of TrustedUpdater to calc_loss() so that it can always the correct
variable value after the checkpoint restoration.