Speed up summary
Currently TrainerConfig.summary_interval controls how often summary is written.
It was expected that the running a large value of summary_interval should be almost
same as debug_summaries=False and summarize_grads_and_vars=False.
But somehow it was not true. In one of my experiments, the difference is as large as 30%.
This PR partially solves this problem by forcing TF to compile two versions of _train().

TODO: Check whether on policy training also has this problem.