RingBuffer: using Event.wait with timeout for more efficient waiting.
This change is tricky to make, because unless we use many Event objects
to track every env whether it has_space or has_data, we will never be sure.
So we ended up using two Events for every dequeue and enqueue, which wakes
up waiting processes more often than necessary and let these processes
check for space or data every so often.

This wastes CPU somewhat, but at least we will not be deadlocked.