Better handling of --conf_param
Previously, if the related module has not been imported, the config
name specified through --conf_param will not be found. This make it
very hard to use since the related modules are typically imported
in the config file. This change delay the binding of conf_param
until the related module has been loaded.