Make sure docs are generated for the scripts under alf/bin
Previously, the docs of some of them are not generated. It seems to be caused
by the fact that some of those scripts try to define same flags when the module
is imported. So I moved the flags definitions into a function so that they are
defined only when the script is invoked.