Suport averaged summary over a summary interval
In some situation, the single summary value might be too volitile, so we allow averaged value over a summary interval if average_over_interval=True

And use average_over_interval for summarize_reward. This is helpful if the episode is long and reward is sparse so that we don't need to wait until episode end to see if the agent gets any reward.