Improve error reporting for alf.nest.get_field()
Previously, it is hard to find out what's wrong when get_field() fails.
This change make the error msg clearer by including all paths in the nest and the desired path.

Also improved the error reporting in config_util so that location of the error in config file will be displayed.