Sampled MuZero
Implements Sampled Muzero described in the following paper:

Hubert et. al. "Learning and Planning in Complex Action Spaces" arXiv:2104.06303

There are two main changes:

1. Changing the prior used for MCTS from original probability to 1/num_sampled_actions
2. Expand all the children of the root node immediately. This is described in Appendix A
of the paper: "At the root of the search tree only, we evaluated all sampled actions before
the start of the search and used those to initialise the Q(s; a) quantities in the PUCT formula".
It turns out that this is very important for the performance.
3. Inspired by 2, I also implemented an option to evaluate all the children of the leaf of a
selected searc path.
4. MCTSAlgorithmTest is modified to test different variants.

Test plan:

1. Change mcts_algorithm_test.py to test all variants
2. muzero_pendulum.gin