Simplify network copy
Previously, the default implementation of the network copy does not copy arguments. So when the network is copied, if some of its components are from the arguments, they are not truly copied. This causes many networks have to implement
their own copy().

This change makes the default implementation of copy() copies the arguments, so we that we don't need to implement the copy() in the subclasses.