Fix a bug of Network.copy() for nn.Sequential
The copy() fails for nn.Sequential when keyword arguments are used
to specifying modules. For example, Sequential(module1, b=module2, a=(('input', 'b'), module3), output=('a', 'b'))

Since element_dict is used to contain those keyword arguments, the
algorithm depends on the order of item in the dict. However, previously,
since map_structure() is used to copy and map_structure will sort the keys in dict,
it can result in incorrect order for element_dict

Solution: not to use map_structure to copy.