Correctly handle TimeStep.discount for MuzeroAlgorithm
Previously, the algorithm assumes that TimeStep.discount can be 0 only at
the end of an episode. This change is to remove this assumption. The purpose
is to support terminal on life loss for Atari, where the discount is set
to 0 upon life loss even though the epsiode does not end yet.