Reduce memory footprint for data shuffle.
For off-policy training, the memory for experience can be large, especially when num_updates_per_iter > 1.
Previously, for the shuffling step, the memory requirement is twice the memory of the experience.

After the change, we don't keep the shuffled experience. Instead, we use the shuffling indices to
extract the mini-batch directly to save memory.

Also record time for preprocess