Speed optimization for MuzeroAlgorithm
With detailed profiling, found that data transformer takes a huge amount of time.
It turns out that it is because they are used under CPU device context in
preprocess_experience. This PR change to do data transformer under
cuda context. This change dramatically speed up the training speed.

Additionally, I also added an option to use AMP for inference (Inference during MCTS
is the major computation load for MuZero) . It can further speed up the training.