Support fractional unroll_length (#1200)
* Support fractional unroll_length and num_updates_per_train_iter

Sometimes, we want to use a large value for num_updates_per_train_iter. But memory
is not large enough for that. With this change, we can achieve the same
effect by using unroll_length < 1.

* Remove supporting of fractional num_updates_per_train_iter

The use case for farctional num_updates_per_train_iter is not convincing.
Only support fractional unroll_length.