Make sure the summaries for unroll are generated according to summary_interval
The period of performing unroll may not be an integer divider of config.summary_interval if config.unroll_length is not an interger. This can lead to the situation that the summaries for unroll are generated at longer periord of up to summary_interval/unroll_length. (if 1/unroll_length is co-prime to summary_interval)

To solve this problem, we use a variable to remember whether a summary is generated between two unrolls and use that to create a new record_if context.