Handle a corner case of summarize_variables (#1376)
Sometimes we might have -inf variables. For example, the log beta is a variable in VAE because sometimes we want to learn it. In other cases we'd like to set it to -inf for a fixed zero beta. So we need to check variable values before summarizing them with a histogram.