Support parallel environments of batched environments
For the previous version of `FastparallelEnviroment`, each process only holds one
environment. So when `num_parallel_environments` is large, we need to create
a lot of processes, which may requires a lot of resources.

In the PR, we allow each process to hold more environments through
``BatchedEnvironmentWrapper``.

To use this feature, simply passing batch_size_per_env>1 for ``utils.create_environment``