Fix config_util for fail to detect same file for import_config()
If import_config() encounters a config file which has been imported,
it will skip it. This behaviour depends on being able to correctly
identify two files are same or not. Due multiple ways of specifying
a same file, this check does not work reliably before. Hence we use
os.path.realpth to canofiy path names to make the check correct.