Fix temporally_independent_train_step (#1518)
Since TrainerConfig is passed to some sub algorithms, `TrainerConfig.temporally_independent_train_step`
can be changed by the sub-algorithm from `None` to False if the sub-algorithm is stateless.
However, the parent algorithm may be stateful. But since config.temporally_independent_train_step
is not `None` anymore, it will not be changed to True even if the parent algorithm is stateful.

Clearly, we want temporally_independent_train_step to be True if the parent algorithm is stateful.