Patch nn.Module.add_module() and do not del Module.__getattr__ (#1522)
1. The new add_module will explicitly set module attributes and hence speedup its access.
2. If all the attributes are explicitly set, nn.Module.__getattr__ will not be called.
So add it back to make things more robust. Note that adding it back slow down attribute access a little bit (~20%),
but considering we already get the most of the speedup and attribute accesss generally doesn't
take too much of overall time (at most a few percent), this should be ok.