sync_progress_to_envs (#1561)
ALF schedulers need to have progress information
to calculate scheduled values. For parallel environments, the
environments are in different processes and the progress information
needs to be synced with the main in order to use schedulers in
the environment. So we add this option so that scheduler can be
used in environment.