Use running statistics to normalize advantages
For some tasks, there are rare but importance events which may not
appear in one batch. In these situations, the advantage normalization based
on one single batch may not correctly normalize the advantage using
the desired statistics. This PR changes the normalization based on
the EMA statistics of the past batches.

Also add support normalize the weighted sum of the advantage instead
of normalization for individual reward dimensions.