Stop async evaluation immediately for Evaluator.close()
When error happens in the training, ALF will try to close all the evaluator.
However, some evaluation takes long time. This cause the main process just
wait and we cannot see the error message from exception.

With this change, the evaluation will stop immediately so we can
see the error messages.