Use CUDA_VISIBLE_DEVICES instead of orch.cuda.set_device to set device for DDP worker
We use CUDA_VISIBLE_DEVICES instead torch.cuda.set_device(rank)
in the training_worker to specify the GPU because we want to
actually only make the specified GPU visible to the process.
This is because some other user of cuda (e.g. Taichi) may
only use the first GPU. If multiple GPUs are visible to the
process, all the processes will compete for the first GPU.