Fix cuda memory allocation issue caused by fused_linear_act.py
In the previous implementation, fused_linear_act.StaticState will
always allocate a cuda tensor once it is imported. The simple act
of allocating a small tensor will cause torch to allocate several
hundred MB cuda memory. This can become very bad if there are a
lot of subprocesses.

Fix is simple, only create the tensor when it is needed.